+++ api/tsconfig.json
@@ -26,7 +26,8 @@
         }
     },
     "include": [
-        "./src/**/*"
+        "./src/**/*",
+        "tests"
     ],
     "exclude": [
         "./node_modules",
